cartoon_diary/requirements/base.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
celery>=5.3,<5.4
redis>=5.0,<6.0
Pillow>=10.0,<11.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
fastapi>=0.111,<0.112
PyJWT>=2.9,<3.0
passlib[bcrypt]>=1.7,<1.8
python-multipart>=0.0.9,<0.0.10
email-validator>=2.2,<3.0
